Renaming files and folders works
Able to rename files and folders by renaming their corresponding inputs
when updating the database. Encountered bug that causes app to crash if
two collections or tracks have the same name and then are deleted one
after another (the first delete removes the collection or track, and
the second delete cannot find it because they were reading the same
files that had been overwritten) Need to add a check to prohibit
creating collections with the same name or tracks within the collection
with the same name.